add forcely sgmllib.py
If your feedparser download included a copy of sgmllib.py, it’s probably
called sgmllib3.py, and you can simply rename the file to sgmllib.py. It
will not be automatically installed using the command above, so you will
have to manually copy it to somewhere in your Python path.